Allow building on Mac or with specific GCC toolchains
The existing makefiles assume a compiler and linker named "g++" and an executable extension of ".exe". To allow testing on Mac or with toolchains named differently (like "g++-7" for instance), this PR allows overloading of the $(LD), $(CXX), and $(EXE) make variables from the command line but otherwise defaults to the same g++/Windows build.

It also adds a  "clean" target to the top-level makefile.