Add Any verification specifier
A flexible test may want to ignore invocations of trivial methods, like
getters and read only functions. Such methods may be called any number
of times, including none.

Currently FakeIt has no way to discard the invocations of a sequence of
methods. Using `ClearInvocationHistory` would discard all the methods,
trivial or not.

This new verification specifier, `Any`, marks a method or sequence
of methods as verified, regardless of the number of invocations
(including none).

Example:

    Verify(Method(mock, important_method)).Exactly(3);
    Verify(Method(mock, trivial_getter)).Any();
    VerifyNoOtherInvocations(mock);

Using this modifier, a flexible test may white-list the trivial methods,
and make sure that important methods are not called.